Project/jarvis/requirements.txt
Python 3.8+
speechrecognition
pyttsx3
requests
python-dotenv